JQ_BE/requirements.txt
 Django>=4.2
 djangorestframework>=3.14
 django-cors-headers>=4.3
 psycopg2-binary>=2.9
 dj-database-url>=2.1
